UiElement and derivated classes now takes a vector of string for classes, instead of a pointer to an array of classes, added tests for style applied to elements

+++ src/element_style/element_style.cpp
@@ -104,7 +104,7 @@ namespace gui::elementStyle {
         return (defaultStyle != nullptr);
     }
 
-    int ElementStyle::nbRules() const { return _rules.size(); }
+    size_t ElementStyle::nbRules() const { return _rules.size(); }
 
     std::string ElementStyle::debugValue() {
         std::stringstream stringStream;
